+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.1 (in development)
 
+- Added tenant scope to `Entry::findSiblings()`
 - Fixed `tenant_id` validation for strings